flex-algo: share FlexAlgoConfig staging engine (parameterized path)
Move the FlexAlgoConfig config-staging engine out of isis/flex_algo.rs
into crate::flex_algo::config, with the config path prefix supplied at
construction. This is the last protocol-neutral piece OSPF needs before
it can stage /router/ospf/flex-algo definitions; the isis-packet wire
builders (build_fad_subs, build_link_asla, ...) and the IS-IS callback
shims stay put.

- crate::flex_algo::config::FlexAlgoConfig::new(prefix) builds the
  per-leaf handler table under {prefix}; exec/commit unchanged.
- isis re-exports FlexAlgoConfig and constructs it with
  "/router/isis/flex-algo" — no behavior change.
- The pure staging unit tests move to the shared module (now exercised
  with a "/router/ospf/flex-algo" prefix to prove genericity).

No behavior change: full zebra-rs suite (769) + yang_load guard pass;
workspace clippy + fmt clean.
